+++ src/main/java/com/backGroundLocate/entity/Leave.java
@@ -9,9 +9,9 @@ public class Leave {
     private int userId;
     private String userName;
     private int type;
-    private int startTime;
-    private int endTime;
-    private int timestamp;
+    private long startTime;
+    private long endTime;
+    private long timestamp;
     private String approver;
     private int approverState;
     private int approverTime;
